Converter now triggers on `lang` attribute for string datatype values
CSVW schemas should use `lang` (not `@lang` or `language`) to indicate the language of the value in a column/row. The `@lang` attribute is JSON-LD internal, and should only be used to annotate the column specification itself. 

The fix involved changing the `csvw.py` file to check for `csvw:lang` rather than `csvw:language`. This was a bug.

This closes #16