removed duplicate folder 'cow', small changes to namespaces
The folder 'cow' was removed since it was an out-of-date duplicate from 'src'.
datalegend.org has been changed to datalegend.net.
Delimiter is now only based on the header, not the entire file.
(published in dev branch to make sure that the removal of the cow folder does not brake anything)